Add style constants, fix dirs, start with property list
