+++ src/constants/theme.js
@@ -1,0 +1,11 @@
+export const COLORS = {
+  background: '#f6efe8',
+  primary: '#611f69'
+}
+
+export const LOGO = {
+  width: 100,
+  height: 100,
+  alignSelf: "center",
+  marginBottom: 30
+}
