+++ src/screens/private/tabs/Tabs.js
@@ -2,12 +2,8 @@ import React, { useState, useEffect } from 'react'
 import { View, Image, TouchableOpacity } from 'react-native'
 import { createBottomTabNavigator, BottomTapBar } from '@react-navigation/bottom-tabs'
 import AsyncStorage from '@react-native-async-storage/async-storage';
-import Home from './hometabs/Home'
-import Profile from './hometabs/Profile'
-import axios from 'axios'
-import Settings from './hometabs/Settings'
-import { REACT_APP_SERVER_URL } from '@env'
-import ProfileHome from './hometabs/misc/ProfileHome';
+import Home from './home/Home';
+import ProfileHome from './profile/ProfileHome';
 
 const Tab = createBottomTabNavigator()
 
@@ -15,19 +11,20 @@ export default function Tabs ({ setIsLoggedIn }) {
   
   
   return (
-    <Tab.Navigator tabBarOptions={{
-      showLabel: false,
-      style: {
-        backgroundColor: 'transparent',
-        borderTopWidth: 0
-      },
-    }} 
+    <Tab.Navigator 
+      tabBarOptions={{
+        showLabel: false,
+        style: {
+          backgroundColor: 'transparent',
+          borderTopWidth: 0
+        },
+      }} 
     >
       <Tab.Screen name="Home"
         options={{
           tabBarIcon: ({ focused }) => (
             <Image
-              source={require('../../../assets/home-black-shape.png')} 
+              source={require('../../../../assets/home-black-shape.png')} 
               resizeMode="contain"
               style={{
                 width: 25,
@@ -45,7 +42,7 @@ export default function Tabs ({ setIsLoggedIn }) {
         options={{
           tabBarIcon: ({ focused }) => (
             <Image 
-            source={require('../../../assets/user.png')} 
+            source={require('../../../../assets/user.png')} 
             resizeMode="contain"
             style={{
               width: 25,
