+++ src/screens/private/tabs/home/styles.js
