+++ src/screens/public/index.js
@@ -1,5 +1,3 @@
 export { default as Signup } from './signup/Signup'
 export { default as Login } from './login/Login'
-export { default as Home } from './home/hometabs/Home'
-export { default as Tabs } from './home/Tabs'
