+++ src/screens/public/login/styles.js
@@ -1,37 +1,35 @@
 import { StyleSheet } from 'react-native';
+import { COLORS, LOGO } from '../../../constants/theme'
 
 export default StyleSheet.create({
     container: {
         flex: 1,
-        alignItems: 'center'
+        alignItems: 'center',
+        justifyContent: 'flex-start',
+        paddingTop: '35%',
+        backgroundColor: COLORS.background
     },
     title: {
 
     },
-    logo: {
-        flex: 1,
-        height: 120,
-        width: 90,
-        alignSelf: "center",
-        margin: 30
-    },
+    logo: LOGO,
     input: {
         height: 48,
         borderRadius: 5,
         overflow: 'hidden',
         backgroundColor: 'white',
+        width: '85%',
         marginTop: 10,
         marginBottom: 10,
-        marginLeft: 30,
-        marginRight: 30,
         paddingLeft: 16
     },
     button: {
-        backgroundColor: '#788eec',
+        backgroundColor: COLORS.primary,
         marginLeft: 30,
         marginRight: 30,
         marginTop: 20,
         height: 48,
+        width: '85%',
         borderRadius: 5,
         alignItems: "center",
         justifyContent: 'center'
@@ -42,7 +40,6 @@ export default StyleSheet.create({
         fontWeight: "bold"
     },
     footerView: {
-        flex: 1,
         alignItems: "center",
         marginTop: 20
     },
@@ -51,7 +48,7 @@ export default StyleSheet.create({
         color: '#2e2e2d'
     },
     footerLink: {
-        color: "#788eec",
+        color: "#611f69",
         fontWeight: "bold",
         fontSize: 16
     }
