--- src/screens/signup/Signup.js
@@ -1,104 +1,0 @@
-import React, { useState } from 'react'
-import { Image, Text, TextInput, TouchableOpacity, View } from 'react-native'
-import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scroll-view';
-import AsyncStorage from '@react-native-async-storage/async-storage';
-import styles from './styles';
-import axios from 'axios';
-import { REACT_APP_SERVER_URL } from '@env'
-
-export default function RegistrationScreen({ navigation, setIsLoggedIn }) {
-  const [fullName, setFullName] = useState('')
-  const [email, setEmail] = useState('')
-  const [password, setPassword] = useState('')
-  const [confirmPassword, setConfirmPassword] = useState('')
-  
-  const onFooterLinkPress = () => {
-      navigation.navigate('Login')
-  }
-
-  const onRegisterPress = async () => {
-    if (password !== confirmPassword) {
-      alert("Passwords don't match.")
-      return
-    }
-    try  {
-      const user = {
-        fullName,
-        email,
-        password,
-      }
-
-      const { data } = await axios({
-        method: 'POST',
-        baseURL: REACT_APP_SERVER_URL,
-        url: `/user/signup`,
-        data: user,
-      })
-
-      await AsyncStorage.setItem('token', data.token)
-      setIsLoggedIn(true)
-    } catch ({ response: { data } }) {
-      alert(`${data.message}`)
-    }
-  }
-
-  return (
-    <View style={styles.container}>
-      <KeyboardAwareScrollView
-        style={{ flex: 1, width: '100%' }}
-        keyboardShouldPersistTaps="always">
-        <Image
-          style={styles.logo}
-          source={require('../../../assets/icon.png')}
-        />
-        <TextInput
-          style={styles.input}
-          placeholder='Full Name'
-          placeholderTextColor="#aaaaaa"
-          onChangeText={(text) => setFullName(text)}
-          value={fullName}
-          underlineColorAndroid="transparent"
-          autoCapitalize="none"
-        />
-        <TextInput
-          style={styles.input}
-          placeholder='E-mail'
-          placeholderTextColor="#aaaaaa"
-          onChangeText={(text) => setEmail(text)}
-          value={email}
-          underlineColorAndroid="transparent"
-          autoCapitalize="none"
-        />
-        <TextInput
-          style={styles.input}
-          placeholderTextColor="#aaaaaa"
-          secureTextEntry
-          placeholder='Password'
-          onChangeText={(text) => setPassword(text)}
-          value={password}
-          underlineColorAndroid="transparent"
-          autoCapitalize="none"
-        />
-        <TextInput
-          style={styles.input}
-          placeholderTextColor="#aaaaaa"
-          secureTextEntry
-          placeholder='Confirm Password'
-          onChangeText={(text) => setConfirmPassword(text)}
-          value={confirmPassword}
-          underlineColorAndroid="transparent"
-          autoCapitalize="none"
-        />
-        <TouchableOpacity
-          style={styles.button}
-          onPress={() => onRegisterPress()}
-        >
-          <Text style={styles.buttonTitle}>Create account</Text>
-        </TouchableOpacity>
-        <View style={styles.footerView}>
-          <Text style={styles.footerText}>Already got an account? <Text onPress={onFooterLinkPress} style={styles.footerLink}>Log in</Text></Text>
-        </View>
-      </KeyboardAwareScrollView>
-    </View>
-  )
-}
